LanggraphProject/router-langgraph/pyproject.toml
[project]
name = "router-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.5.1",
    "python-dotenv>=1.1.1",
]
